Added code to make test case of 1 pass
Summary:
--------

We've added the simplest code (returning 'I') to make this test pass.
Because the test that previously failed now passes we are confident that our code
addresses the test case we've added. If we didn't see the test initially fail we'd
have no confidence that our test and the code actually worked.

Next step:
----------

Write a new test that asserts given 2 the converter should return 'II'.
Again this test should fail.