Added failing test for given 2 we're expecting 'II'
Summary:
--------

We've added a second test. As expected this fails.
There is a bit of duplication in the test code but we'll address that in a few commits time.

Next step:
----------

Write the simplest possible code to make both tests pass.
Again don't do anything clever. We'll improve the solution once we've got more test cases.