Added fix for test case 2
Summary:
--------

Added a quick (and dirty) fix to make all the tests pass. It's still early days in
trying to work out the best solution here so we'll wait till there are more data
points before coming up with a more general solution.

Next step:
----------

Next step is a REFACTOR step. Either production code or test code can be refactored (but not both).
We've got a lot of duplication in the test code.
Write a custom assertion that takes an expected output and an input value. It gets the actual value
from the converter and checks it against the expected value.