Added code to make test case 5 pass
Summary:
-------

The quickest possible solution to make the tests pass has been added.

Next step:
---------

There is no obvious refactor to come up with a more general solution.
However there is lots of repetition in the test code so it's time to tidy that up.
Look up Data Providers in PHPUnit docs.