Added test for 7
Summary:
--------

We've added a test for 7 and didn't get a failing test.
This is because the solution we have is general enough to cope with the number 7.
If you added tests for 8 that too would work straight away.

Next step:
----------

The next problem number will be 9, so write a failing test for this.