Further refactor to remove special case handling of 4 and 9
Summary:
--------

Our refactoring has cleaned up the code even more. It feels like we are getting
closer to a  general solution that can be extended.

Next step:
---------

So far we've dealt with the case where only integers have been supplied.
If the converter is called with a non integer then it should throw an
IllegalAgumentException. Update the PHPdoc to reflect this.